Refactor games repository collection name
Changed the collection name in the games repository from "games_test" to "games" for development use.

Update game result display

Updated the game result display in the GameResultTop widget to show the meter distance from the answer instead of a fixed value.

Adjust minimum score in current game view model

Adjusted the minimum score in the CurrentGameViewModel from 50 to 0 to reflect the score at the outer boundary of the circle.

Calculate distance and score in current game view model

Added code to calculate the distance and score based on the current location and target waypoint in the CurrentGameViewModel.